[REF] website_mass_mailing: convert newsletter subscription Public Widget to Interaction
This widget doesn't need to be enabled in edit mode since it had no effect
there. Indeed, the initialization code was skipped and the subscribe button
wasn't clickable in edit mode.

task-4654032
